+++ tests/RedirectMiddlewareTest.php
@@ -6,6 +6,7 @@ namespace GuzzleHttp\Tests;
 
 use GuzzleHttp\Client;
 use GuzzleHttp\Exception\BadResponseException;
+use GuzzleHttp\Exception\InvalidArgumentException;
 use GuzzleHttp\Exception\ResponseException;
 use GuzzleHttp\Exception\TooManyRedirectsException;
 use GuzzleHttp\Handler\MockHandler;
@@ -81,13 +82,25 @@ class RedirectMiddlewareTest extends TestCase
         $stack = new HandlerStack($mock);
         $stack->push(Middleware::redirect());
 
-        $this->expectException(\GuzzleHttp\Exception\InvalidArgumentException::class);
+        $this->expectException(InvalidArgumentException::class);
         $this->expectExceptionMessage('idn_conversion must be true, false, null, or an integer IDNA_* bitmask');
 
         $stack->resolve()(new Request('GET', 'http://example.com'), [
             'allow_redirects' => ['max' => 2],
             'idn_conversion' => '0',
         ])->wait();
+    }
+
+    public function testRejectsInvalidAllowRedirectsOptionDirectly(): void
+    {
+        $middleware = new RedirectMiddleware(new MockHandler([new Response(200)]));
+
+        $this->expectException(InvalidArgumentException::class);
+        $this->expectExceptionMessage('allow_redirects must be true, false, or array');
+
+        $middleware(new Request('GET', 'http://example.com'), [
+            'allow_redirects' => 'yes',
+        ]);
     }
 
     public function testRedirectsWithRelativeUri(): void
@@ -271,7 +284,7 @@ class RedirectMiddlewareTest extends TestCase
         });
         $request = new Request('POST', 'http://example.com/', [], 'payload');
 
-        $this->expectException(\InvalidArgumentException::class);
+        $this->expectException(InvalidArgumentException::class);
         $this->expectExceptionMessage('stream_factory must be an instance of Psr\\Http\\Message\\StreamFactoryInterface');
 
         $redirectMiddleware->modifyRequest($request, [
@@ -484,7 +497,7 @@ class RedirectMiddlewareTest extends TestCase
         $handler = $stack->resolve();
         $request = new Request('GET', 'http://example.com');
 
-        $this->expectException(\InvalidArgumentException::class);
+        $this->expectException(InvalidArgumentException::class);
         $this->expectExceptionMessage('uri_factory must be an instance of Psr\\Http\\Message\\UriFactoryInterface');
 
         $handler($request, [